+++ verify_Procedure/diffusion term/getGlobalID.m
@@ -1,4 +1,4 @@
 function [globalID] = getGlobalID(~, ysize, xID, yID)
     %>从0开始
-    globalID = int8((xID - 1) * ysize + yID);
+    globalID = (xID - 1) * ysize + yID;
 end
